+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 - Updated `Interfaces/GCClassic/main.F90` so that diagnostic archival is done once per diagnostic timestep (i.e. the larger of the chemistry timestep or dynamic timestep)
+- Updated routine `History_Read_Collection_Data` to set the default update frequency of time-averaged collections to the diagnostic timestep instead of the dynamic timestep
 
 ## [14.7.0] - 2026-02-05
 ### Added